(MAINT) Take snapshot within scenarios before first step is run
This commit adds logic to take a snapshot before the first step of each
scenario from `run-scenario-body-over-steps` is run.  This is done to
ensure that the memory measurement taken for the first step in the
scenario doesn't inappropriately consider memory that may have been
allocated after the step at the end of a prior scenario, between
scenarios, or during the setup phase before the first step of a new
scenario.