(SERVER-1390) Moved get-catalog into a util helper
This commit moves the logic for getting a catalog out from the
`single-catalog-compile` namespace into a new helper function in the
util namespace, `get-catalog`.

The commit also adds a little extra logic to validate that one class
expected in the catalog response is present - to help with ensuring that
the expected catalog is being compiled.  If the class is not present, an
exception is thrown.