(MAINT) Pulled environment-timeout logic into scenario runner
This commit pulls the logic for configuring an `environment-timeout` at
the beginning of a scenario into the common scenario runner code.  This
should reduce the code duplication for different environment-timeout
settings.  The `single-catalog-compile` scenario was reduced to a single
function that is invoked with different environment-timeout values from
the scenario list in `mem-run!`.