(MAINT) Rework for running scenarios in separate JVM instances
This commit reworks the existing scenarios and framework to allow for
the scenarios to be run to be specified via command-line arguments to
the Java process.  This commit also includes an updated bash script
which triggers scenarios in separate JVM processes.  This separation is
done to hopefully reduce the likelihood of memory reductions from
objects allocated in one scenario inadvertently appearing during the
execution of a downstream scenario, when not specifically wanted.
The refactor also increases the likelihood of being able to add on
some scenario combinations without automatically having to change
configuration in Clojure code.