(SERVER-1392) Rename single-catalog-compile and add mult-compile to run
This commit renames the single-catalog-compile scenario to
catalog-one-node-one-jruby-one-environment, in anticipation of new
catalog compilation scenarios to be developed.

This commit adds the multiple-nodes-one-jruby-one-environment scenario
to the default ./dev/run.sh script.

The file names for the baseline, final, and results files created for
each scenario are also now prefixed with the scenario name.