(SERVER-1413) Externalize node name / class configuration
This commit changes the "node" command line parameter to allow for node
name and the corresponding class to verify in catalog compilation output
to be specified as command-line arguments.  This reduces the hardcoding
of information a bit more from the tool source code and makes it a
little easier to run the tool for compilations of nodes not defines in
the perf_control repo.

This commit also reworks the logic in the run.sh script to repeat the
common set of scenarios to run for each node in a reusable function.
This cuts down on a bit of the repetition of scenarios for different
node names which existed in prior commits.

This commit also refactors the scenario step APIs to allow more of the
parameters commonly used by scenarios - step-base-name,
mem-output-run-dir, and jruby-puppet-config - to be passed into the step
callbacks by default.  The prior commit required the desired info to
be closed over in the step callback.

This commit changes the `set-env-timeout!` function to update the
`environment_timeout` in the base environment's "environment.conf" file
instead of the master conf dir's "puppet.conf" file.  This helps ensure
that scenarios override the setting as appropriate, regardless of the
original state of the "puppet.conf" and "environment.conf" files.

For use with puppetlabs-module catalog compilation, a couple of new
facts are added into each catalog request:

- concat_basedir - A custom fact used with the concat module to select
  the base directory into which concat resources are written when
  applied.

- whereami - Custom fact used in puppetlabs-modules to specify location
  of calling agent.