(MAINT) Added group-by-jruby for single catalog to run script
This commit adds a group-by-jruby scenario for a single catalog per
environment/jruby to the run.sh script.

This commit also renames the prior `catalog-one-jruby-one-environment`
scenario to `catalog-group-by-catalog` to distinguish how the memory
snapshots are being grouped in the results - as compared to other
catalog scenarios which group by environment or jruby rather than down
to the individual catalog compiles.