(MAINT) Change method for getting base name of directory for summarize
This method changes the approach used to get the base name of a
directory from using `me.raynes.fs/name` to `.getName` from the
`java.io.File` class.  `me.raynes.fs/name` splits on "." characters when
finding the last part of a name path, which causes an unintended name to
be obtained when the path has a "." character in the name.